// Agilibo.Web/ClientApp/app/models/ClientCompany.ts
export interface ClientCompany {
    companyKey: number;
    companyName: string;
    address: string;
    zipCode: string;
    city: string;
    vatNumber: string;
    countryKey: string;
    countryCode: string;
}

export interface DetailCompany {
    companyKey: number;
    companyName: string;
    address: string;
    zipCode: string;
    city: string;
    vatNumber: string;
    countryKey: string;
    countryCode: string;
    countryName: string;
}